Move walls just a pixel behind the paddle.
So that we don't run into the bug where the ball will go behind the
paddle, scoring 100s of points as it zig-zags between paddle and wall.

And stops the bug where a paddle hitting the ball with the side of the
paddle causes 100s of consecutive saves as the paddle "bounes" off the
ball many times as it moves through the ball.